Updating email to a reusable service instead of a standalone
