+++ src/notification/services/__init__.py
@@ -1,0 +1,5 @@
+from notification.services.email_notification_service import EmailNotificationService
+
+__all__ = [
+    "EmailNotificationService",
+]
